refactor: move parseToken to BearerToken
